modified: various-darwin.zsh; modified: .shared.sh; modified: .zshenv; modified: gen.org; modified: brish.zsh; modified: cleaners.zsh; modified: nightshift.zsh; modified: network.zsh; modified: tags.zsh; modified: text2img.zsh; modified: tmux.zsh; modified: v2ray.zsh; modified: basic.zsh; modified:  ...

+++ scripts/zshlang/auto-load/others/cleaners.zsh
@@ -47,7 +47,9 @@ function rm-caches {
     trs-rm "$(brew --cache)"
 
     pip cache remove '*'
-    conda clean --all --yes #: STILL does not delete old python version stuff!
+    if isdefined conda ; then
+        conda clean --all --yes #: STILL does not delete old python version stuff!
+    fi
 
     trash-empty-all
 }
